[#2145667] Fixed project_release_node_presave() to preserve version info.
In cases where the version subfields are already known and/or don't
necessarily match the version format string, we don't want
project_release to be parsing the full version string and trying to
set the sub elements to the values it thinks are right. For example,
this allows versioncontrol_release to parse the version elements from
a given VCS label and set those itself.